api/requirements.txt
fastapi
uvicorn
openai
python-dotenv
fastapi-cors
pydantic